+++ src/app/(app)/journal/[id]/edit/page.tsx
@@ -45,7 +45,7 @@ export default async function EditEntryPage({
   };
 
   return (
-    <div className="h-full flex flex-col">
+    <div className="h-full flex flex-col animate-in fade-in slide-in-from-right-4 duration-250">
       <EntryForm initial={initial} />
     </div>
   );